Use FromDynamic in more places
